% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_logs_score.R
\name{crps_logs_score}
\alias{crps_logs_score}
\title{crps_logs_score}
\usage{
crps_logs_score(forecast, target)
}
\arguments{
\item{forecast}{a forecast data.frame in long EFI-standard format}

\item{target}{a target data.frame in long EFI-standard format}
}
\description{
Compute the CRPS and LOGS score given a forecast in either ensemble or
normal distribution. (Support for additional distributions to come.)
}
