% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_s3_inventory.R
\name{efi_update_inventory}
\alias{efi_update_inventory}
\title{helper method to run an update of EFI inventory}
\usage{
efi_update_inventory(
  bucket_name = "neon4cast-forecasts",
  endpoint = "data.ecoforecast.org"
)
}
\arguments{
\item{bucket_name}{name of one the EFI buckets.
\code{neon4cast-forecasts} is the most important one to inventory and is actively
used in scoring.  Other buckets can easily open at the repository root.}

\item{endpoint}{endpoint address, default is good.}
}
\description{
helper method to run an update of EFI inventory
}
