% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scores.R
\name{fill_scores}
\alias{fill_scores}
\title{fill scores based on a null}
\usage{
fill_scores(df, null_model = "EFInull")
}
\arguments{
\item{df}{a data frame of forecasts, with column
"team" identifying different forecasts.}

\item{null_model}{the "team" name identifying the baseline (null) forecast
used for filling missing values.}
}
\description{
Comparing mean scores across teams requires appropriate
treatment of missing values: teams should not be able to improve
scores merely by refusing to provide predictions of, e.g. sites
or times which are hardest to predict.  To avoid this,
merely removing missing values when averaging across scores is not
sufficient.  A simple expedient is to replace missing values with
predictions made from a baseline 'null' forecast.  This function
simply provides this behavior.  Original forecast scores with missing values
are retained as \code{crps_model} and \code{logs_model} columns, while
\code{crps} and \code{logs} become filled with baseline scores from the null forecast.
}
\details{
Note that this fills \emph{implicit} NAs, e.g. site/time/variables predicted
by the null team but not predicted by the focal team, as well as explicit
\code{NA}s in the focal team (where the focal team includes each of teams named
in the teams column of \code{df}).
If teams have scores for site/time/variables not forecasted by the "null" team,
these rows are removed and thus cannot contribute to the mean score either.
If the "null" team contains explicit \code{NA}s in scores,
(usually but not always due to missing observations), these are not removed.
}
