% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_horizon.R
\name{include_horizon}
\alias{include_horizon}
\title{include_horizon}
\usage{
include_horizon(
  df,
  allow_difftime = getOption("neon4cast.allow_difftime", FALSE)
)
}
\arguments{
\item{df}{data.frame containing an efi-compliant forecast or score}

\item{allow_difftime}{should \code{horizon} and \code{interval} be expressed as integers (in seconds)
(default) or as native difftime (compatible in R but not in serializations like csv or parquet)}
}
\description{
Adds horizon if not present, defined as \code{time - start_time} with a group
defined by any of: "target_id", "model_id", "pub_time", "variable", "site_id".
Will also add \code{start_time} if not present, defined as \code{time - min(time)}
}
