% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{pivot_forecast}
\alias{pivot_forecast}
\title{pivot forecast}
\usage{
pivot_forecast(df, target_vars = "")
}
\arguments{
\item{df}{standardized target data.frame in wide format}

\item{target_vars}{a character vector of target variable names}
}
\description{
pivot forecast
}
