% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{pivot_target}
\alias{pivot_target}
\title{pivot target}
\usage{
pivot_target(df, target_vars = "")
}
\arguments{
\item{df}{standardized target data.frame in wide format}

\item{target_vars}{a character vector of target variable names}
}
\description{
pivot target
}
