% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score}
\alias{score}
\title{Compute the CRPS & logs score of your forecast}
\usage{
score(forecast, target, ...)
}
\arguments{
\item{forecast}{forecast data frame or file}

\item{target}{target path or URL (csv file)}

\item{...}{additional parameters (Not used)}
}
\description{
Compute the CRPS & logs score of your forecast
}
