% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_forecast.R
\name{standardize_target}
\alias{standardize_target}
\title{Transform older (v0.4) standard to current standard}
\usage{
standardize_target(df, filename = NULL)
}
\arguments{
\item{df}{a forecast data.frame}

\item{filename}{filename, which might be parsed to extract reference_datetime and model_id. Optional.}
}
\description{
Transform older (v0.4) standard to current standard
}
