% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_logs_score.R
\name{summarize_forecast}
\alias{summarize_forecast}
\title{summarize_forecast}
\usage{
summarize_forecast(forecast, extra_groups = NULL)
}
\arguments{
\item{forecast}{a forecast data.frame in long EFI-standard format}

\item{extra_groups}{character vector of additional groups to use for grouping summaries}
}
\description{
Compute summaries of a forecast for ensemble and parametric forecasts
}
