% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score}
\alias{score}
\title{Compute the CRPS score of your forecast}
\usage{
score(
  forecast,
  target,
  theme = c("aquatics", "beetles", "phenology", "terrestrial_30min",
    "terrestrial_daily", "ticks"),
  target_vars = c("oxygen", "temperature", "richness", "abundance", "nee", "le",
    "vswc", "gcc_90", "rcc_90", "ixodes_scapularis", "amblyomma_americanum",
    "Amblyomma americanum")
)
}
\arguments{
\item{forecast}{forecast data frame or file}

\item{target}{target path or URL (csv file)}

\item{theme}{a theme name, an additional metadata field
which is attached as an extra column if not available.  Serves as a forecast series identifier}

\item{target_vars}{for 'wide' format, character vector of valid variable names}
}
\description{
Compute the CRPS score of your forecast
}
