% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_workflow.R
\name{score_theme}
\alias{score_theme}
\title{score_theme}
\usage{
score_theme(
  theme,
  s3_forecasts,
  s3_targets,
  s3_scores,
  s3_prov,
  local_prov = paste0(theme, "-scoring-prov.csv")
)
}
\arguments{
\item{theme}{which theme should be scored?}

\item{s3_forecasts}{a connection from \link[arrow:s3_bucket]{arrow::s3_bucket}}

\item{s3_targets}{a connection from \link[arrow:s3_bucket]{arrow::s3_bucket}}

\item{s3_scores}{a connection from \link[arrow:s3_bucket]{arrow::s3_bucket} where scores will be written.
This connection requires write access, e.g. by specifying}

\item{s3_prov}{a connection from \link[arrow:s3_bucket]{arrow::s3_bucket}}

\item{local_prov}{path to local csv file which will be used to
store provenance until theme is finished scoring.}
}
\description{
A helper utility to score a collection of forecasts efficiently
for challenges hosting submitted forecasts on S3 buckets.
Scores are automatically streamed to an "scores" bucket in parquet format.
A provenance bucket is used to allow function to skip forecast-target
combinations that have already been scored.
}
