% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_forecast.R
\name{standardize_forecast}
\alias{standardize_forecast}
\title{Transform older (v0.4) standard to current standard}
\usage{
standardize_forecast(
  df,
  filename = NULL,
  reference_datetime_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{df}{a forecast data.frame}

\item{filename}{filename, which might be parsed to extract reference_datetime and model_id. Optional.}

\item{reference_datetime_format, }{date or datetime format of reference_time, default to "\%Y-\%m-\%d"
model_id and reference_datetime may be omitted if they are supplied in the
filename.  If these columns exist, then filename is ignored.}
}
\description{
Transform older (v0.4) standard to current standard
}
\details{
Current standard should have columns:
\itemize{
\item model_id
\item reference_datetime
\item site_id
\item variable
\item datetime
\item family
\item parameter
\item prediction
}

This function does not handle un-pivoted (v0.3) forecast, see pivot_forecast()
}
